feat: useNotification hook 구현 및 적용
